+++ CHANGELOG.md
@@ -10,6 +10,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 - Increased `prune` unit test coverage
 
+### Fixed
+- Removed line breaks from `prune`, `reset`, `root` and `up` commands' help messages
+
 ## [0.2.0] - 2026-04-16
 
 ### Added